[StyleCop] Fix all the warnings left - Reorder methods - Reorder usings - Create new files for classes - Reorder usings - Fix encoding
