+++ .NET/Microsoft.Recognizers.Text.DataDrivenTests/Number/TestDecimalAndThousandsSeparators.cs
@@ -2,7 +2,6 @@
 
 namespace Microsoft.Recognizers.Text.Number.Tests
 {
-
     [TestClass]
     public class TestDecimalAndThousandsSeparators
     {
