+++ .NET/Microsoft.Recognizers.Text.DataDrivenTests/TestContextExtensions.cs
@@ -14,9 +14,9 @@ using Microsoft.Recognizers.Text.NumberWithUnit;
 using Microsoft.Recognizers.Text.Sequence;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using DateObject = System.DateTime;
+
 namespace Microsoft.Recognizers.Text.DataDrivenTests
 {
-
     public static class TestContextExtensions
     {
         private static IDictionary<Models, Func<TestModel, string, IList<ModelResult>>> modelFunctions = new Dictionary<Models, Func<TestModel, string, IList<ModelResult>>>()
